Resolvers are refactored now and are stored in separate files. Added updateUser mutation definition and resolver
