+++ src/resolvers/Comment.js
@@ -1,0 +1,14 @@
+const Comment = {  //1:1 // 1:1
+    author(parent, args, { db }, info) {
+        return db.users.find((user) => {
+            return user.id === parent.author
+        })
+    }, 
+    post(parent, args, { db }, info) {
+        return db.posts.find((post) => {
+            return post.id === parent.post
+        })
+    }
+}
+
+export { Comment as default }
