+++ src/resolvers/User.js
@@ -1,0 +1,14 @@
+const User = {     // 1:N // 1:N
+    posts(parent, args, { db }, info) {
+        return db.posts.filter((post) => {
+            return post.author === parent.id
+        })
+    },
+    comments(parent, args, { db }, info) {
+        return db.comments.filter((comment) => {
+            return comment.author === parent.id
+        })
+    }
+}
+
+export { User as default }
