+++ src/schema.graphql
@@ -6,6 +6,7 @@ type Query {
 
 type Mutation {
     createUser(data: CreateUserInput!): User!
+    updateUser(id: ID!, data: UpdateUserInput!): User!
     createPost(data: CreatePostInput!): Post!
     createComment(data: CreateCommentInput!): Comment!
     deleteUser(id: ID!): User!
@@ -16,6 +17,12 @@ type Mutation {
 input CreateUserInput {
     name: String!
     email: String!
+    age: Int
+}
+
+input UpdateUserInput {
+    name: String
+    email: String
     age: Int
 }
